backend/requirements.txt
fastapi
uvicorn
pydantic
pennylane
numpy
google-genai
